seata-order-service mapper 获取返回主键

+++ seata-order-service2001/src/main/java/com/zkc/springcloud/alibaba/dao/OrderDao.java
@@ -8,7 +8,7 @@ import org.apache.ibatis.annotations.Param;
 @Mapper
 public interface OrderDao {
     //1 新建订单
-    Long create(Order order);
+    Integer create(Order order);
 
     //2 修改订单状态 从0 改为1
     void update(@Param("orderId") Long orderId);
